+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Back to development.
+### Added
+- Value groups can use the `flatten` option to indicate values in a slice should
+  be provided individually rather than providing the slice itself. See package
+  documentation for details.
 
 ## [1.10.0] - 2019-11-20
 ### Added